change!: start changing core `Unilateral` class:
Simplify and modularize the basic initialization & some quality-of-life
properties. It should be more readable, more pythonic, and easier to
extend and adapt now. However, the changes to this class are not yet
complete, and it is not in a working state.

BREAKING CHANGE: Code written for previous versions does not run with
this implementation.
Related: #40, #41, #44, #45, #46