change(mods)!: reimplement modalities, add feats
The new implementation should be more flexible, safer, and add more
freedom for defining "weird" diagnostic modalities in case of the
trinary model.

At this point, it is sparsely tested and not even used correctly by the
Unilateral class.

BREAKING CHANGE: This is not compatible with the previous implementation
anymore.