change: move graph representation out of models
Everything to do with Edge, Tumor, and LymphNodeLevel instances should
now be self-contained in the graph.py file and be managed by a
Representation class. Aditionally, I have implemented a simple
trigger-and-callback management that allows me to separate concerns
better: E.g., the Edge instances now don't need to worry about the
transition matrix anymore. When their params are changed, they simply
trigger some callback functions (which should contain the deleter of the
transition matrix).

All of this is still untested.

Related: #44