fix(graph): guarantee order of connections
Previously, I iterated over a set of the connection names that are the
values of the graph dict. This lead to random orders of the edges in the
edge list of the graph representation. In turn, this caused basically
all important matrices to be not deterministic. Thus, the likelihood was
wrong.