change(graph)!: remove edge_params lookup
Instead of meticulously setting up the `edge_params` dictionary with
access to the getter and setter methods, I just provide all nodes and
edges of the graph representation as dictionaries and handle parameter
assignment via that.

BREAKING CHANGE: `edge_params` of the `graph` attribute in the
`Unilateral` class does not exist anymore.

Related: #46