change!: align edge's and dist's get/set params
BREAKING CHANGE: The parameter names are now the edge's or dist's name,
then an underscore and the "kind" of parameter. E.g., for the egdes,
this might be "growth", "spread", or "micro", while for the dists, it
could be any parameter name.
Also, the `get_getters` and `get_setters` has been removed.