fix(bi): delegted generator attr now resets
Previously, the `t_stages` property of the bilateral model was delegated
from the ipsilateral model such that the generator would only yield the
T-stages when first accessed, but remain in that state forever after.
Now, the delegated `t_stages` generator resets every time it is accessed
in the bilateral model.