refactor: `cached_property` over descriptors
Instead of using the convoluted descriptors for stuff like the
transition matrix and the dictionary of diagnose time distributions, I
switched to simple subclasses of the `cached_property` decorator. This
is way shorter, more readable and actually seems to be much safer.

Also, this hopefully allows me to easily share attributes of the
`Unilateral` model with the `Bilateral` instance.