change!: change llh API
The `data` and `load_data_kwargs` are removed from the `likelihood()`
method arguments. The most common use case is anyways to preload the
patient data. This applies to both the uni- and bilateral llh.

In turn, an argument `for_t_stage` was added. It may be used to specify
one particular T-stage for which to compute the likelihood. If it is
omitted, the llh is computed as normally.

This commit also fixes an error in the Bayesian network implementation.
Previously, it called the wrong diagnose matrices. Now, it correctly
computes the BN llh if no T-stage is specified and only that of the
given T-stage if it _is_ provided.

BREAKING CHANGE: `data` and `load_data_kwargs` removed from
`likelihood()` method.