change(diag)!: use composite for distributions
Instead of a dict that holds the T-stages and corresponding
distributions over diagnose times, this implements them as a composite
pattern. This replaces the dict-like API entirely with methods. This has
several advantages:
1. It is more explicit and thus more readable
2. The composite pattern is designed to work naturally with tree-like
structures, which we have here when dealing with bilateral models.
3. This separates a lot of the logic into separate abstract classes,
further separating concerns.

BREAKING CHANGES: The old API of setting and getting distributions will
not work anymore.