feat(type): add model ABC to inherit from
I added an abstract base class from which all model-like classes should
inherit. It defines all the methods that need to be present in a model.

The idea behind this is that any subclass of this can be part of a
composite that correctly delegates getting/setting parameters,
diagnose time distributions, and modalities.