change(uni): compute matrices in data getter
The data and diagnose matrix is now computed on demand in the getter for
the `patient_data` property. This makes sure that the user always sees
an up to date version of the data encoding & the diagnose probabilities.

Also, apparently `del dataframe[column]` is much slower than
`dataframe.drop(columns)`. I replaced the former with the latter and now
the tests are fast again.

Related: #77